backend/requirements.txt
amqp==5.1.1
asgiref==3.2.10
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.14
django-filter==2.4.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
drf-spectacular==0.22.1
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.6.0
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.16
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
